requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.16
httpx>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
google-genai>=1.0.0
python-dotenv>=1.0.0
